// types/patient.ts
export interface Patient {
  PatientID: string
  PatientName: string
  Age: string
  Gender: string
  HomeAddress: string
  City: string
  County: string
  State: string
  ZIPCode: string
  ContactInfo: string
  WeightKg: string
  Race: string
  IncidentNumber: string
  ServiceRequested: string
  OtherAgencies: string
  PrimaryRole: string
  ResponseMode: string
  EMSShift: string
  DispatchCity: string
  DispatchState: string
  DispatchZIP: string
  DispatchCounty: string
  SceneType: string
  Category: string
  BackInService: string
  CrewMembers: string
  NumberOfCrew: string
  OtherAgencyOnScene: string
  NumberOfPatients: string
  PatientContactMade: string
  ArrivedOnScene: string
  FirstOnScene: string
  StagePriorToContact: string
  PrimaryComplaint: string
  Duration: string
  TimeUnits: string
  AlcoholDrugUse: string
  InitialAcuity: string
  CardiacArrest: string
  PossibleInjury: string
  BaseContactMade: string
  SignsOfAbuse: string
  "5150Hold": string
  PastMedicalHistory: string
  CurrentMedications: string
  MedicationAllergies: string
  AdvanceDirectives: string
  HeartRate: string
  BloodPressure: string
  RespiratoryRate: string
  SPO2: string
  Temperature: string
  Glucose: string
  GCS_Eye: string
  GCS_Verbal: string
  GCS_Motor: string
  GCS_Score: string
  GCS_Qualifier: string
  MentalStatus: string
  AbdomenExam: string
  ChestExam: string
  LungExam: string
  SkinAssessment: string
  BackSpineExam: string
  ExtremitiesExam: string
  EyeExam_Bilateral: string
  EyeExam_Left: string
  EyeExam_Right: string
  Medication: string
  Dosage: string
  MedUnits: string
  Route: string
  MedTime: string
  MedResponse: string
  MedComplications: string
  Procedure: string
  ProcLocation: string
  IVLocation: string
  Size: string
  ProcTime: string
  Attempts: string
  Successful: string
  ProcResponse: string
  TransportDisposition: string
  LevelOfCareProvided: string
  TransportAgency: string
  TransportUnit: string
  TransportReason: string
  EMSPrimaryCareProvider: string
  Severity: string
  PrimarySymptom: string
  OtherSymptoms: string
  PrimaryImpression: string
} 